cpu fuser cleanup
