+++ torch/csrc/jit/fusers/cpu/fusion_compiler.h
@@ -3,6 +3,10 @@
 
 #include "torch/csrc/jit/fusers/fuser_interface.h"
 
+#include "torch/csrc/jit/fusers/cpu/tensor_desc.h"
+#include "torch/csrc/jit/fusers/cpu/annotated_graph.h"
+#include "torch/csrc/jit/fusers/cpu/concat_desc.h"
+
 #include "torch/csrc/jit/ir.h"
 #include "torch/csrc/utils/disallow_copy.h"
 #include "torch/csrc/jit/assertions.h"
@@ -15,82 +19,6 @@
 #include <vector>
 
 namespace torch { namespace jit { namespace cpufuser {
-
-
-// type information needed by the compiler for input/outputs
-// contiguity[i] is true if the dim i is contiguous with dim i + 1.
-// contiguity.back() == true means strides.back() == 1.
-struct TensorDesc {
-  at::ScalarType scalar_type;
-  std::vector<bool> contiguity;
-
-  TensorDesc(const at::ScalarType& type, const std::vector<bool>& contiguity)
-  : scalar_type(type), contiguity(contiguity) {
-    nDim_ = std::count(contiguity.begin(), contiguity.end(), false) + (lastIsContiguous() ? 1 : 0);
-  }
-
-  TensorDesc(
-    const at::ScalarType& type
-  , const at::IntList& sizes
-  , const at::IntList& strides)
-  : TensorDesc(type, TensorDesc::findContiguous(sizes, strides)) {}
-  
-  TensorDesc(const at::Tensor& t)
-  : TensorDesc(t.type().scalarType(), t.sizes(), t.strides()) {}
-  
-  TensorDesc(TensorTypePtr type)
-  : TensorDesc(type->scalarType(), type->sizes(), type->strides()) {}
-
-  // number of dimensions after contiguity compression
-  size_t nDim() const {
-    return nDim_;
-  }
-
-  // do we have inner stride == 1?
-  bool lastIsContiguous() const {
-    return contiguity.size() == 0 || contiguity.back();
-  }
-
-  static std::vector<bool> findContiguous(
-    const at::IntList& sizes,
-    const at::IntList& strides);
-
-private:
-  size_t nDim_;
-};
-
-struct AnnotatedGraph {
-  // short-term storage only, so it borrows Graph.
-  AnnotatedGraph(Graph& graph, int device)
-  : graph(&graph), device(device) {}
-
-  Graph* graph = nullptr;
-  int device = kCPUDevice;
-  std::vector<TensorDesc> input_desc;
-  std::vector<TensorDesc> output_desc;
-};
-
-struct ConcatDesc {
-  size_t nSubtensors; // == 1 for outputs that are not concats, otherwise it is the number tensors concatenated
-  size_t dim; // dimension along which the concat occurs
-  std::unique_ptr<TensorDesc> subtensorDesc; // descriptor for the subtensor, if it exists
-  ConcatDesc()
-  : nSubtensors(1), dim(0) {}
-  ConcatDesc(const TensorDesc& desc, size_t nSubtensors, size_t dim)
-  : nSubtensors(nSubtensors), dim(dim) {
-    JIT_ASSERT(nSubtensors > 1);
-    std::vector<bool> cont = desc.contiguity;
-    if(dim > 0) {
-      // when we narrow the concatenated output
-      // we make the size[dim] smaller while keeping the stride[dim] the same,
-      // meaning: stride[dim - 1] != stride[dim]*size[dim]
-      // so dim - 1 is no longer contiguous
-      cont[dim - 1] = false;
-    }
-    subtensorDesc.reset(new TensorDesc(desc.scalar_type, cont));
-  }
-};
-
 
 struct CompiledCPUFusionFunction : public CompiledFusionFunction {
   TH_DISALLOW_COPY_AND_ASSIGN(CompiledCPUFusionFunction);
@@ -185,7 +113,6 @@ private:
 };
 
 CPUFusionCompiler& getCompiler();
-
 
 } // namespace cpufuser
 } // namespace jit
